Add separate range and automatic values to colors
Add the ability to specify different available and automatic ranges for
the coloring provided by DataColorOptions. This is needed to provide
reasonable behavior for the depth map view, where the data range may
change from frame to frame, but, since we don't want to alter user
selected range values, the available range is taken as the union of all
ranges for the data loaded to date.

Also, update DepthMapViewOptions to use this feature.